@trepca 2023-08-06
nice, would be good to also compare memory usage