@arvidn 2023-08-10
I updated the benchmarks with the current version of this patch